+++ app/components/Sidebar/utils.tsx
@@ -1,4 +1,4 @@
-import { FileEntry, readTextFile } from "@tauri-apps/api/fs";
+import { renameFile } from "@tauri-apps/api/fs";
 import {
   SiC,
   SiGo,
@@ -11,15 +11,12 @@ import {
   SiTypescript,
 } from "react-icons/si";
 import { TiDocument } from "react-icons/ti";
-export const Activate = async (file: FileEntry, setState: any) => {
-  const splitted: string[] | undefined = file.name?.split(".");
-  setState({
-    name: file.name,
-    path: file.path,
-    suffix: splitted![splitted!?.length - 1],
-    children: file.children,
-    content: await readTextFile(file.path),
-  });
+
+export const Rename = async (path: string, Name: string) => {
+  let Path = path.split("/");
+  Path.pop();
+  const newName = `${Path.join("/")}/${Name}`;
+  await renameFile(path, newName);
 };
 
 export const LanguageLogo = ({ suffix }: { suffix: string }) => {
@@ -40,7 +37,7 @@ export const LanguageLogo = ({ suffix }: { suffix: string }) => {
       return <SiPython />;
     case "c":
       return <SiC />;
-    case "png": // add more later
+    case "png":
       return <SiImagej />;
     default:
       return <TiDocument />;