+++ app/interfaces/index.ts
@@ -8,14 +8,23 @@ export interface ActiveFile {
   content: string;
 }
 
-export interface LanguageSupport {
-  symbols?: string[];
-  stringLits?: string[];
-  keywords?: string[];
-}
-
 export interface FileComponentProps {
   file: FileEntry;
-  setState: (value: ActiveFile) => void;
   depth: number;
+}
+
+export interface ActiveStore {
+  activeFiles: ActiveFile[];
+  resetActiveFiles: () => void;
+  setActiveFiles: (file: ActiveFile) => void;
+  addActiveFile: (file: ActiveFile) => void;
+  removeActiveFile: (file: ActiveFile) => void;
+}
+
+export interface OpenStore {
+  OpenFiles: FileEntry[];
+  resetOpenFiles: () => void;
+  setOpenFiles: (files: FileEntry[]) => void;
+  addOpenFile: (file: FileEntry) => void;
+  removeOpenFile: (file: FileEntry) => void;
 }