+++ app/lib/modules/html.ts
@@ -1,6 +1,0 @@
-import { LanguageSupport } from "@/app/interfaces";
-
-export const Language_Html: LanguageSupport = {
-  symbols: ["<", ">", "=", "/"],
-  stringLits: ["'", '"'],
-};